change-monitor: wire up diff callback using libgit2-glib.
This seems to mostly get things working using libgit2-glib instead of
the built in monitor for the file.

What we might consider doing next is when opening the file, determine
which version to use. We can keep the old version around for files that
are not yet in a source tree and this around for files that are in a
source tree.

Further refactoring necessary.