vim: a very rough first draft at phrases (dw, yy, gg, etc)
This doesn't support composing them yet. For example, we need to know
if a command should be followed by another (like `d`). Then we know
when we need to wait for more input.

After that is implemented, we need to be able to pass the parameter to
the function to perform that command.

Something like "10dw"

 - 10 is the number of times to perform the operation
 - d is the operation
 - w is the modifier (for word)

This could then be performed like

  gb_editor_vim_do_phrase (vim, 10, lookup (commands, 'd'), 'w')

Lookup() would return a function callback based on the name. Something
like:

 gb_editor_vim_phrase_delete (GbEditorVim *, gchar modifier)