vim: Differentiate between inclusive/exclusive motions
Motions are either inclusive or exclusive. While there is no difference
when used standalone, when the motion is used as modifier to a command,
the range of characters the command applies to differs - in case of an
exclusive motion, the command does not apply to the left-most character
described by the motions.
An example probably illustrates it best - with a buffer of 'foo bar baz',
and the cursor at the beginning of the 2nd word:

 - 'de' deletes 'bar' (as 'e' is an inclusive motion)
 - 'dl' deletes 'b' (as 'l' is exclusive)
 - 'dh' deletes the space before 'bar' (as 'h' is exclusive as well)

The current behavior implements the one for inclusive motions, this
patch flags exclusive motions as such and corrects their behavior.

https://bugzilla.gnome.org/show_bug.cgi?id=737959