vim: Fix selection-bound off-by-one error in forward-word
'w' should move to the start (e.g. first character) of the next word.
The iter we move around indeed corresponds to this position, so the
movements without a selection turns out right - however when using
the iter as selection-bound, the selections extents up to - but not
including - the iter's position. Moving the iter one character further
fixes this, though we temporarily break its use as modifier ('dw') due
to another off-by-one error in the opposite direction.

https://bugzilla.gnome.org/show_bug.cgi?id=737959