vim: Support linewise motions
Motions can be either character- or linewise when used as modifiers
to a command. The commands will operate on entire lines when the
motion is linewise, e.g. 'dj' deletes two lines (rather than from
the cursor position to to the same column on the following line).
Add another flag to support this.

https://bugzilla.gnome.org/show_bug.cgi?id=737959