vim: Generalize support for {visual}commands
Replace the current workaround for yank with a VISUAL flag to mark
commands that can be used with an existing selection. This is still
quite different from the handling in vim, where selections constitute
a separate mode with its own command set - so 'y' in normal mode is
a different command than '{visual}y' (a.k.a. 'v_y') in visual mode.
Usually {visual} commands correspond closely to a normal mode command
using the same key though, so this approach should work just fine
for the most common cases.

https://bugzilla.gnome.org/show_bug.cgi?id=737912