change-monitor: add trailing newline if necessary
GtkSourceBuffer has a feature called implicit-trailing-newline that
will automatically hide the trailing newline from documents. This is
especially useful in .c or .h documents. However, it breaks the change
monitor making it think the trailing newline has been removed.

This simply adds that newline back to the document to its original form
if that feature is enabled.