prefs: start on GbEditorSettingsWidget
This widget will be used to manage the values of GbEditorSettings. When
selecting a language, this will be shown to configure that individual
language's settings.

It appears that this is going to be too big to do in a popover, so we
will probably use another dialog to conifgure them, as annoying as that
may be. At least it somewhat matches the design of Network settings
in the Control Center.