prefs: add GbSourceStyleSchemeButton and GbSourceStyleSchemeWidget
This are convenience classes for selecting a style scheme in a more visual
way than just from a list of scheme names. This shows a bit of a preview
of the style-scheme using C as the language in the buffer.

This could probably use a bunch of work, but I think it's better than just
a list even at this point.