editor: move "auto-indenter" into GbSourceView
This simplifies things a bit more, and prepares stuff in a way that would
be more suitable for inclusion in GtkSourceView.

GbSourceView will maintain an auto_indenter based on the selected
language. However, if the parent auto-indent property is false, no
indentation will be used.

This also ensures the configuration is hooked up to the settings widget.