editor: add grid design to Builder style scheme
Since we can't do this from the style scheme, we just add support for it
in the GbSourceView. It means that other style schemes can't get support
for it, but we'll fix that when we reach that. We can push the feature
into GtkSourceView potentially. Might as well test it here first.