html: add very simple html autocompletion
Someone should take this and run with it. In particular, we should
import all the fancy w3c stuff (or load it dynamically from a dtd).
Additionally, we should import various CSS stuff so we can complete
those properly.

Also, when detecting CSS, we always suggest css key names. We should
determine when we are in a CSS value, and limit the results to values
that are available for that key.

We should do the same for HTML elements by only showing valid attribute
names.