search: implement new search design
             **This is not ready for general consumption**

The new search design was laid out on the Builder mailing list[1]. This
is the beginning of that implementation.

## Pending Work

 * Keyboard navigation is not solved
 * "X More" result navigation is not implemented
 * We need a plaholder page when no search is performed
 * Additional search providers.
 * Connecting activation signals
 * etc ...

## Performance Notes

In particular, the old search felt slow for a couple of reasons.

 * We inserted delays to slow down search which were too large.
   Well at least when the search layer is faster.
 * We created way too many result objects. Since these were also
   GtkWidgets, the overhead was huge.
 * We were adding these widgets and sorting them in GtkListstore,
   which was not meant to handle the load. realize() costs alone
   where too much.

So this adds a new structure called GbSearchReducer. It's purpose is to
help us avoid creating objects if we know they will simply be discarded
based on the expected number of results. Since most search layers will
calculate the score as they walk through the result set, all we need is
the score to determine if we need to inflate a full result object.

This alone makes search much faster when walking through the Git trie.

[1] https://mail.gnome.org/archives/builder-list/2015-January/msg00002.html