widget: add gb_widget_activate_action()
This is a helper to activate an action based on what the action sees as
available actions, working it's way up to the toplevel widget.

If "win" is specified, and no other groups overrode "win", then the
toplevel will be used.

If "app" is specified, and no other groups override "app", then the
default GApplication will be used.