[PATCH] souce-vim: Improve scrolling and clean up
We need to make sure the scroll even for horizontal movements,
so make sure we do that when no vertical alignment is needed.
Also, as a clean up, the iter bound is always depending of the
selection and if it is a selection present. So just calculate that
inside ensure_scroll.

https://bugzilla.gnome.org/show_bug.cgi?id=742855