vim: add basic support for vim splits via ctrl+w operations
This supports creation of vertical and horizontal splits as well as
closing the current split. Note that this only plumbs the signals, and
the embedder is responsible for implementing what a split means.

This does not currently implement support for navigation through the
splits. That can be implemented separately from this.

https://bugzilla.gnome.org/show_bug.cgi?id=743020