editor: restore cursor position when refocusing
Since GtkTextBuffer manages cursor position and selections in the buffer
rather than in the view, we need to fake things out a bit when we move
back and forth.

This still doesn't solve some annoyances such as the current line
highlight being off, or seeing the cursor move when not focused. But
it's a whole lot better than completely losing your previous cursor
postiion.

What we should really do, is for Gtk4 move selections to the GtkTextView
and out of the GtkTextBuffer.