search: Scroll to lefmost keeping word visible
When searching currently we put the searched word in the middle
of the view.
But normally the context of the program is at the left, so it's
interesting to keep the most part of the left side of the code
visible, while keeping the searched word fully visible.

To achieve that scroll the iter to the end of the word, keeping that
iter the rightmost possible, so the view stay in the leftmost possible.