tests: add simple test case for IdeContext
This just tries to load a simple autotools based project and checks that
a few items are there.

We are currently bundling all of the test data, since the tests are
published with the sources. I'm not sure how I feel about that long term,
but for now it will do.