vcs: add some safeguards to DirectoryVcs
Try to not load so many files with directory vcs. It's okay to "soft fail"
after say 2000 files have been loaded. We want to be careful with the
situation where we open files starting from the home directory (which
shouldn't happen after we have project loading, but does now).

Also, the directory loader needs to be rewritten to be synchronous and
run from a GTaskThreadFunc.