vim: support replaying '>' with '.' command
One reasonable way to reindent code in vim is to select
the lines, hit '>' to indent one time and then hit '.'
several times until the desired indention level is reached.
This doesn't work in gnome-builder because the '>' command
isn't recorded.

This commit make sure '>' (and '<') get recorded, so they
can be replayed with '.'

https://bugzilla.gnome.org/show_bug.cgi?id=744286