widget: add gb_widget_bind_context() helper
This helper can be used to track changes in the toplevel workbench or
loaded contexts within the workbench. This is useful so that we can load
and unload contexts as projects are opened or closed.

As we move more features out of GbWorkbench and into IdeContext, this will
become more and more useful.