libide: add IdeUnsavedFile abstraction
This gives us a boxed type for accessing a specific file. It is
immutable to make it easy to pass to worker threads that need to perform
work based on the GBytes content. Such an example is clang.

We track sequence numbers so that we can potentially skip work in the
future if nothing has changed.