libide: start on IdeGcaService
This service will help us manage connecting to GCA and providing proxy
structures. Doing so will allow us to progress towards creating an
IdeGcaDiagnosticsProvider. Once that is in place, we basically have
everything we have today in src/gca, but with the ability to bring some
parts in process (like clang).