libide: add ide_buffer_get_diagnostic_at_iter()
This is a helper function to get a diagnostic that can be found at a
given iter in the source view. This is handy if you want to display a
tooltip.

This is not particularly optimized, but that can come later. Ideally,
we'll have some sort of bloom filter to know if there is a diagnostic
on a line before traversing the structure.