libide: add IdeMakecache
Makecache is cached makefile output that gets mmap()'d into the containing
process. The goal is to reduce the amount of calls to make to generate
intermediate information. Additionally, we want to avoid loading lots of
strings as resident data. So instead, we just mmap() the makecache file
and use GRegex to walk through the content as necessary.

We also aggresively cache the results, to avoid multiple lookups.

Once we know the target<->file mapping, we can use make again to get
and extract the CFLAGS for the target. We aggresively cache this
information too.

This is heavily based upon makeintegration.py from gnome-code-assistance.

TODO:

Invalidate the makecache when the makefile or dependencies change. However,
I think this belongs outside of the IdeMakecache module to simplify this
code to avoid locking annoyances.