libide: use mark for page scrolling
Information about positioning of textiter's is updated in an idle
callback. That means that they wont always be up to date. If we use
scroll_to_mark instead of scroll to iter, it can finish the scroll
after the line information has been calculated.

This fixes spurious scrolling up and down with ctrl+f/b in vim mode.