libide: use alternative join-lines implementation
This wont swallow text after the end of the lines. Not that big of a deal
to pull all the text out and reinsert it since it's wrapped in a
begin/end user action. On large chunks of text, its probably easier on
the buffer too since there are fewer iter changes to propagate.