libide: add macro recordings for common vim mutations
Not everything gets recorded as macros, just specific things. So we need
to go through and plumb these through.

Visuals will still need work as we need to start recording when we
transition, but complete upon a valid mutation. Some follow up
sequences (such as `y`) will need to cancel the capture and rollback to
the previous capture since they do not mutate the buffer.