libide: add IdeSourceView:smart-backspace gproperty
Smart backspace will attempt to backspace extra characters when you are
at the indent boundary. It will mind the tab_indent and indent_width
properties to generally do the right thing.

If you are a tab, and the indent_width is less than the tab width, spaces
will be added until you reach the target indent width (after removing the
tab).