views: add document selection popover
I'm not sure about this long term, but this is a stop gap until we figure
out the design further. In particular, it is difficult to select documents
that are not yet in git.

Some alternate approaches are to simply add the files to git immediately.
This was going to be the case once we have templates for things like
GObjects, widgets, resources, icons, etc. However, that doesn't help the
situation for temporary buffers.

Clearly we could choose a better icon.

This needs design review, and some careful planning as we go forward. I
have a couple other ideas on placement, but this is the minimal patch
that solves the problem at hand.


Signed-off-by: Christian Hergert <christian@hergert.me>