makecache: walk through makecache a line at a time
This allows us to avoid using MULTI_LINE regex (not really a big deal), but
more importantly, it allows us to check for additional types of matches
while walking through the content.

In particular, we need to track the current subdir, which will be doable
now in a single pass through the data.