makecache: add support for recursive automake
The previous makecache code was expecting non-recursive automake (since
Builder was the test case). Of course, there is plenty of automake projects
that use recrusive automake, so we need to support that too.

This adjusts our make calls to be able to execute within the target subdir
so that targets are available.

There is the slight chance for error if the same basename exists in
multiple subdirs. We need to think about the best way to handle that
going forward. Probably doable by choosing from the available targets
to which most closely matches the relative path.