clang: try to reduce concurrent requests into a single
If we have a bunch of requests for the same file, just do the work once
an then notify the others on the way out.

This code is pretty hideous. It's also difficult to debug. What we
*really* need is a nice abstraction that ties together with GTask to
do this type of work.

libiris had this, and *mabye* it makes sense to bring that back. Because
then we could have just done an "iris_task_add_callback()" to propagate
the result to additional tasks.