sourceview: add IdeSourceView::focus-location gsignal
This signal is emitted when the sourceview has determined that it needs
to jump to a new source location. This could happen after resolving
a type that we'd like to jump to, or a multitude of things we haven't
thought about yet.

This is also subject to change, but makes forward progress.