editor: add rudimentary symbol browser
If the IdeLanguage provides an IdeSymbolResolver, allow jumping to a
symbol within the file directly from the drop down.

We have a lot to do here, to make this feel nice, but this is the general
plumbing required. We should add icons for symbol type (class, enum,
struct, func, etc) as well as deal with keyboard navigation better.