libide is a GPL-3.0 shared library
Typically I use LGPL-2.1+ for shared libraries. However, due to the
nature of libide being the "guts" of Builder, I do intend for this to
be GPL-3.0. Additionally, it simplifies what we can include in the
library without worrying about staying LGPL safe.

Exceptions will of course be made for upstreaming into libraries such as
GtkSourceView. Just ask if you want something in particular.