clang: when diagnosing, prefer translation unit from .c over .h
This should work for C++ as well, but has not yet been tested.
Now, working on headers should be more useful since we'll use the
translation unit for the .c file if it exists. This does have the
caveat that your .c file needs to include the header. But it really
isn't much worse than we have today, which is non-optimal.