command-bar: handle focus changes to GtkStack
This will look up the widget hierarchy to see if the focus widget is in
a GtkStack. If so, it will ensure that the focus widget is a child of
the GtkStack:visible-child. If not, it will focus the new visible-child.

It would be nice to handle a few more types that GtkStack long term.