source-view: improve scroll-offset with large value
Setting a scroll offset of 999 should keep the current line pinned to the
center of the screen when possible. We had some strange situations where
we would actually move a little bit which was jarring.

This doesn't get at the root of the issue, but it does mitigate it enough
to be much more useful.