icon: provide a symbolic variant of the app icon
Since 3.16 the high contrast accessible theme is
covered by the recolorable symbolic variant of the app icon.

The symbolic size has been introduced to hicolor to cope with
the nominal size of 16x16 rather than the backward compatible
128x128.

Some of the publically installed icons actually belong to the local
resource too.

https://bugzilla.gnome.org/show_bug.cgi?id=744487