style: add styling to the project tree
This adds some basic styling to the project tree. I don't actually want
it to look like a normal treeview. The fact that it is a treeview is
actually a side-effect of the implementation.

I think soon we will need to break our CSS out into 3 parts. Things that
can be shared across all themes, light, and dark modes. Although I'm not
sure what all can be shared yet.