project-selector: make sure the entry gets the first keypress
I'm not thrilled about doing it this way, since most code in other places
passes the event to the search bar.

But, since we are doing a workaround for the search bar (because of 1px
height constraints), we can just send the event to the search entry.

Ideally we would change this after searchbar doesn't have the 1px height
requirement and we can drop our workaround.