style: make some css styling a bit more theme agnostic
Some of this isn't specific to Adwaita. To get the level of quality we
have put into the theme we support (Adwaita), theme authors are going
to have to spend some time getting the details right in their themes
on a per-application basis.

Builder has a lot of detailed modifications to Adwaita. In particular,
you'll notice the columns in the editor workspace. We highlight things
appropriately based on focus.

You'll also need to determine what entries you want linked for your theme.
We've done a few work-arounds to get the aesthetics we want there as well.

Some stuff, like setting treeview colors needs to be done by theme authors.
We can't get that right for ever theme using named colors.