minimap: use IdeSourceView
This allows us to render line changes in the gutter which can be sort
of useful. Since most of our heavy lifting is done in the IdeBuffer, it
shouldn't be a whole lot more heavy handed than the GtkSourceView
approach. Although, we should keep an eye on things to make sure that
stays the case.