stack: focus previously focused view when closing current view
We tried to do this previously, but the focus list was getting mucked
with between the them we removed the item and the time we checked the
list.

This instead checks the list before performing the operation so that we
have access to it even when the focus chain has been mutated due to
intermittent GtkStack changes.