minimap: use gtksourceview instead of idesourceview
IdeSourceView is probably a bit too heavy to use here. It also makes it
harder for us to share this code with libgtksourceview.

Instead, we will use GtkSourceView here and pack the renderer we want
manually.

When moving this code to GSV, we should make it possible to pack custom
renderers in the source view from the .ui definition.