libide: restore previously loaded files when loading context
This will restore any of the previously loaded files when restoring the
context.

   ** I'm sorry if this causes a bunch of files to open on you **
   ** initially, since we failed to prune them in the past.    **

Once you close them (and wait a few seconds for the reclamation to
happen), they shouldn't come back.

If it gets really annoying, and you want to have a clean slate, just
remove everything in ~/.local/share/gnome-builder/drafts/