project-tree: stash tree paned position after timeout
We were only updating the value when we would save/hide the project tree.
This does it after a second delay when the paned is moved. That should
still handle the animation case, where this will be spuriously updated.

Clearly, this is the type of stuff that belongs in a layout manager
abstraction.