minimap: support GdkEventScroll to scroll source view
This will translate the scroll event into a ::move-viewport on the
primary GtkSourceView widget.

It would be nice if we could deal with smooth scrolling here, but the
X11 gdk backend seems to be doing that pretty well with my touchpad.

We'll see how big of an issue it is based on user feedback.