style: add work around for 10pt Cantarell interface font
The top bar was broken for Cantarell at 10pt font because GtkEntry
was not expanding to the allotted space. (Allocation was correct,
but the draw area was not).

This simply works around the issue by using the same margin we would
get from image-button and applies it directly to margin-top.