gca-service: don't use shared dbus session bus
In this case, we were using the shared dbus connection. This two can be
closed out from under us and cause us to fail to save unsaved files.

Instead, this uses a private dbus connection to the session bus for
GCA related activity.

Additionally, we handle the closed signal and release the connection and
proxies, causing them to be reloaded on the next request.