minimap: animate the minimap in/out of view
When the mouse is placed over an editor widget (such as the source map
or the source view), animate the source map in. When it leaves, animate
it out after a short delay.

I assume we want to add to this the ability to hide the map once typing
begins similar to the scrolled window.