makecache: fix shell parsing for simple makefile lines
For simpler lines, where we don't parse earlier in the line, we could hit
a situation where \ is on the suffix. We don't actually need the additional
line information, and this simply was breaking g_shell_parse_argv().

For now, let's just drop the \ from the line. Longer term, we probably
need to investigate whether or not we need the additional lines. My
5 minute test did not show positive results in doing so.

This happens to fix loading parsing of CFLAGS for the systemd project for
me. However, I have not exhaustively checked.