egg-signal-group: add EggSignalGroup helper
This is a work in progress to simplify connecting to a group of signals
on a particular object that may change during runtime. Things like
signal blocking will be managed and transfered to the new object.

I expect this to go through a bunch of iterations as we use it to
simplify code in modules like ide-source-view.