source-view: use ::bind and ::unbind properties
This will use the ::bind and ::unbind properties on the EggSignalGroup
so that we manage the buffer change as on set.

Also, we don't want to deal with the initial GtkSourceBuffer, so we
transform non IdeBuffer instances to NULL.

Ideally, we would just use GtkTextViewClass::create_buffer(), but we
cannot since IdeBuffer relies on so much within the IdeContext. (We
also require an IdeBuffer::file property, which would never be set).