egg-binding-set: add grouped GBinding helper
EggBindingSet let's you create a group of bindings that you would like
to apply in the future on an, yet unknown, object. Simply call
egg_binding_set_set_source() with the source object for the binding
at some point in the future and the bindings will be applied.

You can change the source at any time to apply the bindings to a new
source instance.