source-view: use EggBindingSet
This uses EggBindingSet to manage the group of bindings that we need to
track using the IdeBuffer:file property. Since there is a layer of
indirection here, it's convenient to manage stuff like this as a group.

Long term, I'd love if we could connect this without even having to use
::bind and ::unbind. A dot notation might be neat. Something like:

  "buffer:file"

That would allow us to track changes to both buffer, and it's file
property to map to the source object. Anyway, all in good time.