editor-map: add border to editor map
This adds a slight border as recommended to us by the design team. We are
simply using GtkSeparator for the line since CSS borders do not seem to
get drawn here.

We might be able to simplify this once we move to GtkSource for the minimap
implementation.

It was recommended that we use slightly lighter color than @borders for
this, which seems to be what is getting drawn. We can adjust with CSS
if necessary.