settings: add IdeSettings for project vs global setings abstraction
We have the need to have a layered approach to settings. We need settings
that first resolve via the project's configured settings, and then from
the user defaults.

We have a couple of ways to approach this long term, but this get's us
started by storing project preferences in a subpath such as:

  /org/gnome/builder/projects/<project-name>/<relative-path>

The global settings for this would be stored at:

  /org/gnome/builder/<relative-path>

When configuring the settings via IdeSettings, we will apply the settings
to the project directly.

There is a lot still missing from here.

For example, it would be nice to choose whether or not to apply the
project settings. Otherwise, we'd need one path for GSettings and another
path for IdeSettings.

I tried writing a custom GSettingsBackend, but that become difficult in
propagating changes from underlying backends. We might be able to get
around this by holding on to GSettings instead of a GSettingsBackend.