file-settings: implement IdeFileSettings with xmacros
All of these settings were getting out of hand. At least now they are all
in one place, making it a bit simpler to add a new one.

Not great, but cleaner to add the various _set gproperties.

In many ways, I wish this was just a GSettings that we could implement with
custom backends (say EditorconfigGSettingsBackend). But that would be
so specific that it seems like overkill.

Suggestions? Send them my way.

The next step here is to allow us to have mutliple file settings
implementations for a file. That way, unset values can fall through to
the next provider. This might happen by making FileSettings more of a
tree (with children implementations as an array in a GPtrArray).