file-settings: lazy initialize GAsyncInitable implementations
This also means we are adding a settled property to discover when the
children file settings have completed their async initable initialization.

This means that we can change other parts of builder to get the
file-settings synchronously, and they will back-fill as I/O can occur.
This should simplify things in source-view which is pretty
troublesome today.