egg: make me a sandwich
A better attempt at layered gsettings.

This time, we use a gsettings with memory backend to cache the current
state of all the layered gsettings. This allows us to use regular
gsetting bindings. We simply invalidate the memory backend state when
changing the settings.

This might be a bit heavy handed, but I'd pretty much do anything to
not have to 1) write a custom gsettings backend. 2) duplicate the
circular protection in g_settings_bind().