egg: add high-performance rdtscp based EggCounter
EggCounter is a performance counter based on ideas from previous work
on high performance counters. They are not guaranteed to be 100%
correct, but they approach that with no synchronization given new
enough hardware. In particular, we use %ecx from rdtscp (the core id)
to determine which cachline to increment the counter within.

Given a counter, the value will be split up int NCPU cachelines where
NCPU is the number of cores returned from get_nprocs() (on Linux).

Updating the counter is very cheap, reading back the counter requires
a volatile read of each cacheline. Again, no correctness is guaranteed.

In practice, very few values are lost even during tight competing loops.
A loss can happen when the thread is pre-empted between the %rdtscp
instruction and the addq increment (on x86_64).

If you are not on x86_64, or are missing the rdtscp instruction, a 64-bit
atomic will be performed using __sync_fetch_and_add8(). Clearly, we can
do some more work here to abstract which implementation is used, but we
only support GCC and Clang today, which both have that intrinsic. Some
architectures may not have it (such as 32-bit PPC), but I'm not too
concerned about that at the moment.

The counters are mapped into a shared memory zone using shm_open() and
mmap(). An external program can then discover the available counters
and print them without blocking the target program. It simply must
perform the reads in a volatile manner just like the target process
would need to do for readback.

EggCounterArena provides a helper to walk through the counters in the
shared memory zone. egg_counter_foreach().

You cannot remove a counter once it has been registered.

To define a counter, you must have support for constructor attributes.

  EGG_DEFINE_COUNTER(Symbol, "Category", "Name", "Description")

To increment the counter in a function of your choice (but within the
same module), use EGG_DEFINE_ADD, EGG_DEFINE_INC(), or EGG_DEFINE_DEC().

The layout of the shared memory zone is broken into "cells". Each cell
is an approximate cacheline (64-bytes) on modern Intel hardware. Indexes
to data locations are represented in cells to simplify the math and
allow the compiler to know we are working with properly aligned structures.

The base pointer in EggCounter.values is not 64-byte aligned! It is 8-byte
aligned and points to the offset within the cacheline for that counter.
We pack 8 64-bit counters into a single cacheline. This allows us to avoid
an extra MOV instruction when incrementing since we only need to perform
the offset from the base pointer.

This comes with a program to dump the counters from the Builder process
called ide-list-counters. Use it as such:

  ./ide-list-counters pid

You might even want to watch the output with watch like:

  watch -d -n 0.5 ./ide-list-counters pid

Happy Hacking!