source-view: improve handling of Control+BackSpace
GtkTextView universally sucks at handling Control+BackSpace. It is
non-obvious where the backspace will end and often jumps multiple lines.

As such, we will override that behavior when smart-backspace is emabled.

If you are at the beginning of a line, ctrl+backspace will simply move
you to the end of the previous line. I believe this is the right
behavior because anything else would require you to be cross-eyed and
looking at the start and end of the line at the same time. Non-obvious.

When there is only whitespace between the insertion cursor and the
beginning of the line, we will remove that space and leave the insert
mark at line offset 0.