search-bar: only handle Esc before propagating
This moves the majority of key handling to be G_CONNECT_AFTER. This means
that widgets like GtkEntry and such do not need special cases anymore.

We still handle Escape before propagation otherwise it will get swallowed.

This switches to EggSignalGroup since that is much more convenient to
connect/disconnect groups of signals. Obviously.