style: don't use GtkSeparator for row separators
This is really slow on gtk master. So while the current situation adds a
bit of a extra styling at the bottom of the list, the 10s of seconds of
reduced startup time is well worth it.

Once more cssnode stuff lands and we can disable it for the last row,
(or first, depending on implementation), things will be better.