makecache: use EggTaskCache for file targets cache
This replaces our custom lock chicanery for positive and negative cache
entries for filename<->maketarget cache.

It looks like about a zero sum in terms of lines of code, but it is
definitely simpler to think about in logic. Everything goes through
the cache, no more crazy locking to deal with and debug.