context: add ide_context_hold() and ide_context_release()
ide_context_hold() will add a hold to the IdeContext preventing it from
being unloaded. It also increments the reference count, but that is an
implementation detail and should not be relied upon.

Unloading is different that dispose/finalizing. Unloading the context
does a series of asynchronous events such as saving modified buffers
and releasing allocations in a safe manner.

Calling ide_context_release() will release the hold count on the context
and allow any previous call to ide_context_unload_async() to continue.