mingw: hold reference during async startup
We were able to get into a situation with the unit tests where this thread
func would not get called until after the context had been disposed. Since
we need access to the context (but IdeObject is now allowed to take a full
reference), we need to have a hold on the context.

Should fix transient test failures.