autotools: track buffer changes for makefile invalidation
Clearly, the right way to do this is to watch all of the Makefile.am and
included makefiles (often .mk) for changes using inotify/kqueue/etc.

However, that is a lot more work and requires real plumbing. Something we
want to do, but after we have time to write real makefile parsers.
Instead, let's just track buffer changes and invalidate the makecache if
something that resembled a makefile was edited by the user.

We wont catch external modifications, but at least internal ones will
work for now.