autotools: simply makecache generation and caching
This does a couple things, because it was difficult to split up the
commits.

 - switches to FINAL type and uses IdeAutotoolsBuildSystem as the private
   struct, avoiding all the private nonsense.
 - Uses EggTaskCache to generate the makecache in a race-free manner.
 - Fixes a situation where we would never dispatch completion of the
   makecache generation.

I know it seems sort of heavy to use EggTaskCache for a single item in
the cache. However, simply having the protection against racing and
dispatching to multiple waiters is reason enough to switch. It already
fixes a really difficult to track down issue where we could fail our
`make -p -n -s` job and not notify workers. This caused the spinner to
spin forever, and never actually make forward progress.