highlight: Seperate GtkTextTag list to private and public
Private tags are tags that are used troughout highlighter update function
and are cleared on each iteration.There is no way to get a hold on a private
tag outside of IdeHighlighterEngine.

Public tags are tags that are used outside of the update scope and are
only cleared on reload.A public tag can be retrieved via ide_highlight_engine_get_style.
example: IdeXmlHighlighter

note: Both private and public tags are kept in sync with GtkSourceStyleScheme