source-view: disable <Return> accelerator from completion window
I've noticed a problem where completion proposals are getting activated
when all you want to do is enter a newline. This results in having to
hit escape a lot just to escape the autocompletion.

Instead, we'll just rely on tab (and accelerators) to complete and make
Return just add a new line.