ctags: background build ctags indexes occasionally
We really need to play with how often we will do this. It might be nice
if we could break them up into multiple files so we don't re-index the
entire project all the time.

Another approach we could do later is an append only ctags file that gets
GC'd occasionally. Then we insert !lines containing enough information
to know the target file and ignore files that are repeated.