recent-projects: add support for removing recent projects
This allows us to remove recent projects from the recent-used.xbel
projects file. We pass a list so that we don't pathalogically open/close
the bookmark file when removing a group of entries.

This will remove disovered projects from the list, but they will still
return the next time the miner discovers them (currently, next startup).