webkit: disable unneeded WebKit features
This is partially a workaround for WebKit bug #143245, where the
inspector process pool and default process pool (the default web
context) access the HTML5 local storage database at the same time from
different threads. In theory, that's an SQLite bug, but it doesn't
really matter because GNOME Builder is not a web browser and it doesn't
need HTML5 local storage. Let's avoid the issue altogether by disabling
it.

This probably won't stop Builder from accessing the local storage
database altogether, because it won't affect the inspector web context,
but now it should only happen from one thread at a time. WebKit master
is smart enough to not create the inspector web context unless it's
actually needed, so going forward that shouldn't be an issue anymore
(since Builder does not enable the web inspector).

Note: it would probably be a bit better to disable these settings before
creating the web view, rather than after. But I'm not sure how to do
that with the web views living in .ui files, and they seem happy there.
There might be a teensy race left open, but it's a WebKit or SQLite bug
rather than a Builder bug, and nobody will ever hit it, knock on wood.

https://bugzilla.gnome.org/show_bug.cgi?id=750163