rg: add realtime scrolling graph implementation
Lightly based on the ideas from writing Uber Graph years ago. However, this
is much more simplistic in that we don't yet do the ring buffer for the
backing pixmap. It makes it much simpler to wrap my head around in the
process (while costing a bit more CPU). Last I checked (a few years ago)
the pixmap saved an extra 5% CPU depending on the number of data points
as well size of the graph. At some point, we will probably add this.

The CPU value parser is rather crappy. sscanf() is generally pertty slow,
so if you feel like optimizing this, go for it. It would be nice to have
it abstracted to support FreeBSD/OpenBSD/etc as well.

We still need to add scales and other necessary components around the
graph to allow them to be interacted with. But it's time to get this
in the primary repository for review.

I assume long term we'll make these visible when we execute the program.
One of the best things I ever did when developing previous software was
ensure I watched CPU/Memory/Network graphs while testing and debugging
code. I'd like to see that here as well.

To begin with, we'll probably make an additional side pane that can be
enabled/disabled with an easter egg for now to play with various graphs.

To have a go at the graph and test how smooth it can render on your system,
use the test application in tests/

 ./test-cpu-graph -m 1 -s 60

The above command will perform 1 sample per second with a visible range of
60 seconds on the graph. The Y axis is from 100.0% to 0.0%.